pyproject.toml
[project]
name = "trello-mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.12.4",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
